requirements.txt
bottube>=1.6.0
python-dotenv>=1.0.1
schedule>=1.2.2
